+++ NEWS.md
@@ -1,6 +1,8 @@
 # cbcTools 0.0.3
 
 - Modified the argument of `cbc_profiles()` to `...` so that the user no longer needs to create a separate list to define the attributes and levels.
+- Modified the arguments for the `randN()` and `randLN()` functions to `mean` and `sd`.
+- Updated names of random parameter models to match that of logitr v0.6.0.
 - Updated documentation and examples for all functions.
 - Adding piping example to readme.