+++ NEWS.md
@@ -2,7 +2,8 @@
 
 - Modified the argument of `cbc_profiles()` to `...` so that the user no longer needs to create a separate list to define the attributes and levels.
 - Modified the arguments for the `randN()` and `randLN()` functions to `mean` and `sd`.
-- Updated names of random parameter models to match that of logitr v0.6.0.
+- Improved printing of counts in `cbc_balance()` and `cbc_overlap()`.
+- Updated names of random parameter models to match that of future logitr v0.6.0.
 - Updated documentation and examples for all functions.
 - Adding piping example to readme.