+++ NEWS.md
@@ -1,3 +1,11 @@
+# cbcTools 0.0.5
+
+- Updates for compatibility with logitr version 0.8.0.
+
+# cbcTools 0.0.4
+
+- Updates for compatibility with logitr version 0.7.0.
+
 # cbcTools 0.0.3
 
 - Modified the argument of `cbc_profiles()` to `...` so that the user no longer needs to create a separate list to define the attributes and levels.